Turn EditionWithdrawer into an edition service
This change makes the 'unwithdraw' action a first-class citizen
in the publishing workflow for editions, and makes it easier to
integrate into the admin interface (as it now behaves like the other
workflow actions).
